cloud/websocket: fix handshake error during cloud test
This patch fixes an issue that occurs during the cloud websocket
test. When SDR is disabled, client certificate and key are still
passed to the SE config. This patch solves the issue by calling
setting the SE parameters only when the '-s' argument is passed.

This patch also adds handling of the new connection error to be
passed to the JS layer.

Change-Id: If64d1a10ab21a833da68419b71db3257d1ab13a8
Signed-off-by: Geoffroy Letourneur <g.letourneur@samsung.com>