examples/gatt: Change battery UUID to a custom UUID
Bluez 5.48 introduces a new battery plugin that wraps the battery
profile and exposes it directly over DBUS. In this case the client
cannot find the battery service in the service list anymore.

This patch changes the UUID to a custom one in order not to interfere
with the battery plugin.

Change-Id: I6425167f18c15874fbfc07cc4e7728982e903434
Signed-off-by: Camille Bégué <c.begue@samsung.com>